# packages/contracts/openapi/feedyarder.openapi.yaml
openapi: 3.1.0
info:
  title: Feedyarder API
  version: 0.1.0
  description: Public API contract for the Feedyarder web client.
servers:
  - url: http://localhost:3001
paths:
...
  /setup:
    post:
      operationId: createInitialUser
      summary: Bootstrap the single local user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: Setup completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Setup already completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /session:
    post:
      operationId: createSession
      summary: Create a login session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSession
      summary: Delete the current session
      responses:
        "204":
          description: Session deleted
  /me:
    get:
      operationId: getCurrentUser
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /folders:
    get:
      operationId: listFolders
      summary: List folders
      responses:
        "200":
          description: Folder list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FolderResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /folders/{id}:
    patch:
      operationId: updateFolder
      summary: Update folder metadata
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFolder
      summary: Delete a folder and leave its feeds unassigned
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Folder deleted
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createFolder
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feeds:
    get:
      operationId: listFeeds
      summary: List feeds
      parameters:
        - in: query
          name: includeStatistics
          required: false
          schema:
            type: boolean
            default: false
          description: Include stored item and read-item counts for each feed.
      responses:
        "200":
          description: Feed list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createFeed
      summary: Create a feed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedRequest"
      responses:
        "201":
          description: Feed created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Feed URL already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feeds/discover:
    post:
      operationId: discoverFeeds
      summary: Discover feeds advertised by a webpage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoverFeedsRequest"
      responses:
        "200":
          description: Feed discovery result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedDiscoveryResponse"
        "400":
          description: Invalid or unsupported URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Webpage cannot be inspected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Webpage fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feeds/{id}:
    patch:
      operationId: updateFeed
      summary: Update feed metadata or paused state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedRequest"
      responses:
        "200":
          description: Feed updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Feed not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Feed URL already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFeed
      summary: Delete a feed and cascade-delete its items and fetch history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Feed deleted
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Feed not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /feeds/{id}/retry:
    post:
      operationId: retryFeedNow
      summary: Queue a feed for immediate retry and ensure it is unpaused
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Feed scheduled for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Feed not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /fetch-events:
    get:
      operationId: listFetchEvents
      summary: List recent fetch events
      parameters:
        - in: query
          name: feedId
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Fetch event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FetchEventResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /opml/import:
    post:
      operationId: importOpml
      summary: Import feeds and folders from an OPML document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportOpmlRequest"
      responses:
        "201":
          description: OPML import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportOpmlResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /opml/export:
    get:
      operationId: exportOpml
      summary: Export the current feed tree as OPML
      responses:
        "200":
          description: OPML document
          content:
            application/xml:
              schema:
                type: string
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items:
    get:
      operationId: listItems
      summary: List items with cursor pagination and filters
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: feedId
          schema:
            type: string
            format: uuid
        - in: query
          name: folderId
          schema:
            type: string
            format: uuid
        - in: query
          name: read
          schema:
            type: boolean
        - in: query
          name: starred
          schema:
            type: boolean
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Item page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{id}/state:
    patch:
      operationId: updateItemState
      summary: Update read/starred state for a single item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemStateRequest"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    SetupRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 256
    SessionRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    UserResponse:
      type: object
      additionalProperties: false
      required: [id, username, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: true
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    CreateFolderRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 128
        position:
          type: integer
          minimum: 0
    UpdateFolderRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 128
        position:
          type: integer
          minimum: 0
    FolderResponse:
      type: object
      additionalProperties: false
      required: [id, title, position, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
    CreateFeedRequest:
      type: object
      additionalProperties: false
      required: [feedUrl]
      properties:
        feedUrl:
          type: string
          format: uri
        authUsername:
          type: string
          minLength: 1
          maxLength: 256
          description: Optional HTTP Basic auth username used only by the worker fetcher.
        authPassword:
          type: string
          minLength: 1
          maxLength: 1024
          description: Optional HTTP Basic auth password. Write-only; never returned by feed responses.
        title:
          type: [string, "null"]
          maxLength: 256
        siteUrl:
          type: [string, "null"]
          format: uri
        folderId:
          type: [string, "null"]
          format: uuid
    DiscoverFeedsRequest:
      type: object
      additionalProperties: false
      required: [url]
      properties:
        url:
          type: string
          format: uri
...
    FeedDiscoveryResponse:
      type: object
      additionalProperties: false
      required: [siteUrl, feeds]
      properties:
        siteUrl:
          type: string
          format: uri
        feeds:
          type: array
          items:
            $ref: "#/components/schemas/DiscoveredFeed"
...
    ImportOpmlRequest:
      type: object
      additionalProperties: false
      required: [opml]
      properties:
        opml:
          type: string
...